backend/requirements.txt
Django==5.1.5
django-environ==0.12.0
psycopg2-binary==2.9.10
djangorestframework==3.15.2
django-filter==24.3
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
django-cors-headers==4.6.0
gunicorn==23.0.0